Split the code in three main scripts
