+++ fic_modules/helper_functions.py
@@ -1,4 +1,11 @@
 import re
+import sys
+import os
+from github import Github, GithubException
+from fic_modules.configuration import (
+    REPOSITORIES,
+    REPO_LIST
+)
 from datetime import datetime
 
 
@@ -166,4 +173,49 @@ def filter_strings(string):
             string = remove_chars(string, word)
 
     return string
+
+
+def limit_checker():
+    """
+    This function checks if your limit requests is not exceeded.
+    Every time when this function is called, it returns 1 in case of your
+    requests limit is not exceeded, otherwise it will wait for the reset time
+    to pass.
+    :return: returns 1 if your limit requests is not exceeded
+    """
+    TOKEN = os.environ.get("GIT_TOKEN")
+    GIT = Github(TOKEN)
+    rate_limit = GIT.rate_limiting[0]
+    unix_reset_time = GIT.rate_limiting_resettime
+    reset_time = datetime.fromtimestamp(unix_reset_time)
+    if rate_limit >= 5:
+        sys.stdout.write("\rRate limit is: " + str(rate_limit))
+        sys.stdout.flush()
+        return True
+    else:
+        try:
+            print("You have reached the requests limit!")
+            print("The requests limit will reset at:" + str(reset_time))
+            while rate_limit < 5000 and reset_time >= datetime.now():
+                unix_reset_time = GIT.rate_limiting_resettime
+                reset_time = datetime.fromtimestamp(unix_reset_time)
+            print("\nThe requests limit has been reset!")
+            return True
+
+        except GithubException.status == 403:
+            print("The requests limit is reset to: " + str(reset_time))
+        except GithubException.status == 404:
+            print("Github is down!\n Please try again later...")
+
+
+def get_keys(name):
+    """
+    :param name:
+    :return:
+    """
+    for key in REPOSITORIES.get("{}".format(name)):
+        REPO_LIST.append(key)
+    print(REPO_LIST)
+    return REPO_LIST
+
