+++ git_files/mozapkpublisher.json
@@ -1,11 +1,11 @@
 {
   "0": {
-    "lastChecked": "2019-02-12 08:01:59.061421",
+    "lastChecked": "2019-02-13 04:53:03.489501",
     "last_releases": {
       "latest_release": {
-        "version": "1.0.0",
-        "sha": "88026ea7e7350a2149427692f9d364fa4b5be12a",
-        "date": "2019-02-11 20:47:22",
+        "version": "1.0.1",
+        "sha": "9e0bb1d38057d78dc463e252790c9a517f54b8ba",
+        "date": "2019-02-13 00:40:01",
         "author": "mitchhentges"
       }
     }
