+++ git_files/mozapkpublisher.md
@@ -1,4 +1,4 @@
-## MOZAPKPUBLISHER COMMIT MARKDOWN TABLE SINCE 2019-01-29 10:14:20.291344
+## MOZAPKPUBLISHER COMMIT MARKDOWN TABLE SINCE 2019-01-30 06:52:56.404797
 
 | Commit Number | Commiter | Commit Message | Commit Url | Date | 
 |:---:|:----:|:----------------------------------:|:------:|:----:|
