+++ hg_files/autoland.md
@@ -1,8 +1,13 @@
 ## REPOSITORY NAME: AUTOLAND
- CURRENT PUSH ID: 77957
+ CURRENT PUSH ID: 78059
 
 | Changeset | Date | Commiter | Commit Message | Commit URL | 
 |:---:|:---:|:----:|:----------------------------------:|:-----:| 
+|78040|2019-02-13 01:00:25|Johan Lorenzo <jlorenzo@mozilla.com>|Bug 1505538 - Release x86_64 Fennec Nightly in the Google Play Store r=mtabaraRelease x86_64 Fennec Nightly in the Google Play StoreDifferential Revision: https://phabricator.services.mozilla.com/D19534|https://hg.mozilla.org/integration/autoland/pushloghtml?changeset=de25d2f1abef
+|78028|2019-02-12 23:53:30|Andreea Pavel <apavel@mozilla.com>|Backed out 3 changesets (bug 1525959, bug 1526002) for failing androidBacked out changeset de0efca1118e (bug 1526002)Backed out changeset 503cbc86e442 (bug 1525959)Backed out changeset 33ea61c54aea (bug 1525959)|https://hg.mozilla.org/integration/autoland/pushloghtml?changeset=34f33ae5913e
+|77999|2019-02-12 18:35:42|Mike Hommey <mh+mozilla@glandium.org>|Bug 1525760 - Followup: re-enable the tasks without sccache. r=meAnd refresh rust dependency after bug 1525733.|https://hg.mozilla.org/integration/autoland/pushloghtml?changeset=8b2e8f177263
+|77999|2019-02-12 18:35:42|Marco Castelluccio <mcastelluccio@mozilla.com>|Bug 1527227 - Update Rust nightly version. r=chmanchester a=Aryx|https://hg.mozilla.org/integration/autoland/pushloghtml?changeset=76639dcdef8f
+|77996|2019-02-12 18:27:35|James Willcox <snorp@snorp.net>|Bug 1525959 - Run GeckoView mochitests on x86_64 emulators in automation r=gbrownDifferential Revision: https://phabricator.services.mozilla.com/D19015|https://hg.mozilla.org/integration/autoland/pushloghtml?changeset=33ea61c54aea
 |77930|2019-02-12 01:36:16|Andrew Halberstadt <ahalberstadt@mozilla.com>|Bug 1519472 - [taskgraph] Factor logic for adding a cache in job.common to a new function, r=tomprinceWe add caches at various places in common.py. This consolidates the logic intoa re-useable function. This is in preparation for adding generic-worker cachesupport.This also adds a test. The test is not terribly useful, but I've been lookingfor an excuse to lay some groundwork for further tests in the 'job' submodule.This will do.Differential Revision: https://phabricator.services.mozilla.com/D17689|https://hg.mozilla.org/integration/autoland/pushloghtml?changeset=036604abf1e5
 |77930|2019-02-12 01:36:16|Andrew Halberstadt <ahalberstadt@mozilla.com>|Bug 1519472 - [ci] Opt out of caching for generic-worker based Windows builds, r=tomprinceThe hosts don't have enough disk space to cache mozilla-central.Differential Revision: https://phabricator.services.mozilla.com/D18738|https://hg.mozilla.org/integration/autoland/pushloghtml?changeset=2053a035eee6
 |77930|2019-02-12 01:36:16|Andrew Halberstadt <ahalberstadt@mozilla.com>|Bug 1519472 - [taskgraph] Support generic-worker caches in run_task, r=tomprinceThis implements support for adding generic-worker caches. As a consequence thisalso turns on caching for the gecko checkout if present.Differential Revision: https://phabricator.services.mozilla.com/D17690|https://hg.mozilla.org/integration/autoland/pushloghtml?changeset=887cc76ba189
