+++ hg_files/mozilla-inbound.md
@@ -1,8 +1,12 @@
 ## REPOSITORY NAME: MOZILLA-INBOUND
- CURRENT PUSH ID: 111862
+ CURRENT PUSH ID: 111889
 
 | Changeset | Date | Commiter | Commit Message | Commit URL | 
 |:---:|:---:|:----:|:----------------------------------:|:-----:| 
+|111889|2019-02-13 03:03:53|Geoff Brown <gbrown@mozilla.com>|Bug 1527011 - Support win64-aarch64 artifact builds; r=nalexander|https://hg.mozilla.org/integration/mozilla-inbound/pushloghtml?changeset=636d2c00234d
+|111875|2019-02-12 18:37:32|Marco Castelluccio <mcastelluccio@mozilla.com>|Bug 1527227 - Update Rust nightly version. r=chmanchester a=Aryx|https://hg.mozilla.org/integration/mozilla-inbound/pushloghtml?changeset=76639dcdef8f
+|111874|2019-02-12 18:24:37|Geoff Brown <gbrown@mozilla.com>|Bug 1525738 - Stop running Android 7.0 x86_64 geckoview-junit tests; r=mbrubeck|https://hg.mozilla.org/integration/mozilla-inbound/pushloghtml?changeset=a90e2dd8b1a5
+|111863|2019-02-12 10:19:15|Mike Hommey <mh+mozilla@glandium.org>|Bug 1525760 - Followup: re-enable the tasks without sccache. r=meAnd refresh rust dependency after bug 1525733.|https://hg.mozilla.org/integration/mozilla-inbound/pushloghtml?changeset=8b2e8f177263
 |111858|2019-02-12 07:32:32|Chris Manchester <cmanchester@mozilla.com>|Bug 1525733 - Require rust 1.32 to build. r=firefox-build-system-reviewers,tedDifferential Revision: https://phabricator.services.mozilla.com/D19244|https://hg.mozilla.org/integration/mozilla-inbound/pushloghtml?changeset=5041739e5a5b
 |111858|2019-02-12 07:32:32|Chris Manchester <cmanchester@mozilla.com>|Bug 1526505 - Update mozharness stats parsing code for per-language stats. r=firefox-build-system-reviewers,tedDifferential Revision: https://phabricator.services.mozilla.com/D19247|https://hg.mozilla.org/integration/mozilla-inbound/pushloghtml?changeset=9cc8cc20982c
 |111858|2019-02-12 07:32:32|Alok Kumar <ialokkumarsingh0@gmail.com>|Bug 1492279 - Only fetch target task for actions that require it r=dustinDifferential Revision: https://phabricator.services.mozilla.com/D19273|https://hg.mozilla.org/integration/mozilla-inbound/pushloghtml?changeset=bb243f4edd88
@@ -2891,8 +2895,8 @@
 |86848|2017-02-12 04:31:25|Madeleine Chercover <madeleinechercover@me.com>|Bug 1338385 - Add cron job for nightly coverage tests run; r=jmaherMozReview-Commit-ID: GiL4LZgHmqW|https://hg.mozilla.org/integration/mozilla-inbound/pushloggerhtml?changeset=55dfa1cb4a75
 |86848|2017-02-12 04:31:25|Sebastian Hengst <archaeopteryx@coole-files.de>|Bug 1338385 - Add cron job for nightly coverage tests run. Follow-up: Use two blank lines after definitions. r=flake8-fix|https://hg.mozilla.org/integration/mozilla-inbound/pushloggerhtml?changeset=16e5b52a9a68
 |86848|2017-02-12 04:31:25|Wes Kocher <wkocher@mozilla.com>|Merge m-c to autoland, a=mergeMozReview-Commit-ID: 4z9MXOzghI7|https://hg.mozilla.org/integration/mozilla-inbound/pushloggerhtml?changeset=779d10ed78f5
-|86848|2017-02-12 04:31:25|Gregory Szorc <gps@mozilla.com>|Bug 1336607 - Remove dummy geckolib and taskgraph filter; r=froydnjNow that we use the real geckolib and have all dependencies vendored,the dummy geckolib is no longer required, so we remove it.Also, the taskgraph code for testing for Servo's presence alwayspasses and is no longer needed, so we remove it.Pushed on a CLOSED TREE because ¯\_( )_/¯MozReview-Commit-ID: ITAqArK4Bks|https://hg.mozilla.org/integration/mozilla-inbound/pushloggerhtml?changeset=4db2b9c21878
-|86848|2017-02-12 04:31:25|Gregory Szorc <gps@mozilla.com>|Bug 1336607 - Fix lint failure on a CLOSED TREE; a=¯\_( )_/¯|https://hg.mozilla.org/integration/mozilla-inbound/pushloggerhtml?changeset=d2ec43037c70
+|86848|2017-02-12 04:31:25|Gregory Szorc <gps@mozilla.com>|Bug 1336607 - Remove dummy geckolib and taskgraph filter; r=froydnjNow that we use the real geckolib and have all dependencies vendored,the dummy geckolib is no longer required, so we remove it.Also, the taskgraph code for testing for Servo's presence alwayspasses and is no longer needed, so we remove it.Pushed on a CLOSED TREE because  \_( )_/ MozReview-Commit-ID: ITAqArK4Bks|https://hg.mozilla.org/integration/mozilla-inbound/pushloggerhtml?changeset=4db2b9c21878
+|86848|2017-02-12 04:31:25|Gregory Szorc <gps@mozilla.com>|Bug 1336607 - Fix lint failure on a CLOSED TREE; a= \_( )_/ |https://hg.mozilla.org/integration/mozilla-inbound/pushloggerhtml?changeset=d2ec43037c70
 |86848|2017-02-12 04:31:25|Phil Ringnalda <philringnalda@gmail.com>|Merge autoland to m-c, a=merge|https://hg.mozilla.org/integration/mozilla-inbound/pushloggerhtml?changeset=43e17d5a1492
 |86848|2017-02-12 04:31:25|Phil Ringnalda <philringnalda@gmail.com>|Merge m-c to m-i|https://hg.mozilla.org/integration/mozilla-inbound/pushloggerhtml?changeset=aa57bb701206
 |86826|2017-02-11 01:33:17|Kim Moir <kmoir@mozilla.com>|Bug 1313389 - ci for quantum graphics r=bustage|https://hg.mozilla.org/integration/mozilla-inbound/pushloggerhtml?changeset=299ee41d08aa
